// node/configure/type/thumb-image.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * 自然言語による、人間が見て分かりやすい名前を設定する。通知メールの件名などで使用される。
 */
export type NoName1 = string;
export type NoName3 = string;
export type NoName4 = string;
export type NoName6 = string;
export type NoName7 = string;

export interface NoName {
  title: NoName1;
  endpoint: NoName2;
}
export interface NoName2 {
  dev: NoName3;
  production: NoName4;
  auth: NoName5;
}
export interface NoName5 {
  username: NoName6;
  password: NoName7;
}